add typescript with babel config + upgrade libs
