+++ .babelrc
@@ -1,5 +1,5 @@
 {
-  "presets": [
+  "presets": [ "@babel/typescript",
     ["@babel/env", {
       "targets": {
         "browsers": [
@@ -11,4 +11,8 @@
       "modules": false
     }]
   ],
+  "plugins": [
+      "@babel/proposal-class-properties",
+      "@babel/proposal-object-rest-spread"
+  ]
 }
