+++ package.json
@@ -2,7 +2,7 @@
   "name": "phaser3-project-template",
   "version": "1.1.0",
   "description": "A Phaser 3 Project Template",
-  "main": "src/index.js",
+  "main": "src/index.ts",
   "scripts": {
     "build": "webpack --config webpack/prod.js ",
     "start": "webpack-dev-server --config webpack/base.js --open"
@@ -21,18 +21,22 @@
   "devDependencies": {
     "@babel/core": "^7.7.2",
     "@babel/preset-env": "^7.7.1",
+    "@babel/preset-typescript": "^7.12.1",
     "babel-loader": "^8.0.6",
     "clean-webpack-plugin": "^3.0.0",
-    "file-loader": "^4.2.0",
+    "file-loader": "^4.3.0",
     "html-webpack-plugin": "^3.2.0",
     "raw-loader": "^3.1.0",
-    "terser-webpack-plugin": "^2.2.1",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0",
+    "source-map-loader": "^1.1.1",
+    "terser-webpack-plugin": "^2.3.8",
+    "ts-loader": "^8.0.6",
+    "typescript": "^4.0.3",
+    "webpack": "^4.44.2",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^4.2.2"
   },
   "dependencies": {
-    "phaser": "^3.20.1"
+    "phaser": "^3.24.1"
   }
 }
