+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+        "noImplicitAny": true,
+        "module": "es6",
+        "target": "es5",
+        "allowJs": true,
+        "moduleResolution": "node",
+        "allowSyntheticDefaultImports": true
+        
+    },
+    "include": [
+        "./src/**/*", "index.d.ts"
+    ]
+}
